chore(pkg)!: Move packages under pkg for reuse
BREAKING CHANGE: packages are now in the pkg folder

- in preparation of merging Release-notary and Commitsar git-tool packages it's important to prepare folder structure
- packages currently in release-notary will be moved here as this is the main project
