+++ pkg/history/commit_text.go
